requirements.txt
Django==2.0.4
psycopg2
psycopg2-binary
whitenoise
social-auth-app-django
PyGithub
